[Requirement #9885] - Perform design reviews
Ensure all solutions build:

- Fixed ImageViewerSamples.sln

+++ ImageViewer/Tools/ImageProcessing/RoiAnalysis/PathProfileComponent.cs
@@ -125,7 +125,7 @@ namespace ClearCanvas.ImageViewer.Tools.ImageProcessing.RoiAnalysis
 			{
 				int rawPixelValue = image.PixelData.GetPixel(pixel.X, pixel.Y);
 				_pixelIndices[i] = i;
-				_pixelValues[i] = image.ModalityLut[rawPixelValue];
+				_pixelValues[i] = (int) image.ModalityLut[rawPixelValue];
 				i++;
 			}
